backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.17.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
reportlab==4.2.2
sqlparse==0.5.5
tzdata==2025.3
